Do not load unnecessary script files
